backend/requirements.txt
django==3.0
djangorestframework
django-cors-headers